TAG v2.0.1.30 main changes:     36477 <multimedia> <feat> add cst816d tp driver code     39360 <multimedia> <perf> optimize the lcd tp i2c init     40413 <multimedia> <feat> add new spi lcd driver code     40414 <multimedia> <feat> add spi lcd demo project     37248 <tools> <feat> Add repo manifest version info when generate libs     39041 <platform> <fix> Idle task assert V2     39039 <platform> <fix> API that rtos_get_time_diff may overflow     39050 <platform> <perf> It cost too time time when call bk_aon_rtc_get_us in ISR     37512 <wifi> <feat> Add fast dhcp with dhcp-request way     40388 <ble><fix> use rpa when first adv start     40559 <pm> <fix>fix reboot failure issue     40176 <pm> <fix>fix iperf deauth issue of iperf project after ps close cmd     40857 <bt> <feat>support configuring directed advertising     40724 <platform> <fix> Turn on the DMA copy function of USB     40725 <platform> <perf> Increase the judgment of whether the detection device is fully released when a disconnection occurs     40705 <platform> <fix> reset reason lost if crash in ISR     40715 <bt> <docs>optimize bluetooth project doc     40145 <wifi> <perf> optimize web vnd_cal and saradc instructions     40171 <platform> <perf> optimize multi adc,add CCA interface     40174 <wifi> <perf> optimize phy log,11AX PER,RF RU test     40175 <wifi> <perf> add RF RU test     40703 <wifi> <perf> optimize rssi accuracy     37295 <wifi> <perf> write mac addr to otp1     41079 <wifi> <feat>add license to https     40361 <wifi> <perf> [RF] add crc verify when save RF calibration in OTP     41172 <bt> <feat>add scan stopped event     37778 <pm> <fix> pm debug cli cause SecureFault     41171 <wifi> <perf> optimize xtal bound checking     41077 <wifi> <fix> Stack mem overflow happened with IPV6 while save fast connect info     41176 <platform> <fix> fix bl2 compilation error when TFM_BL2_LOG_LEVEL = NONE     41291 <pm> <fix>add HW low voltage wakeup delay delta to wifi tbtt delay     41260 <wifi> <fix> change timer to oneshot timer in rwnx_txq     41411 <multimedia> <feat> add spi lcd display with qspi module     41532 <bt> <fix>add processing for failure to get feature     41707 <pm> <fix>fix  current increasement issue after second entry into low voltage state with wifi disabled     41535 <platform> <feat> gpio retention support     41526 <wifi> <feat> support country code 2.0     41689 <pm> <perf>modify the low voltage from 0.5v to 0.6v     41758 <multimedia> <docs> improve project and document     41713 <bt> <perf>optimize ble throughput result display     41114 <platform> <fix> Exit from low-voltage,task watchdog timeout assert     41786 <multimedia> <fix> i2s test case work error     41710 <wifi> <fix> [RF] avoid RFPLL unlock when dual-85 test     41775 <wifi> <fix> [RF] fix bad EVM of some boards when low temperature     41776 <wifi> <fix> [RF] fix WIFIPLL amplitude control issue     37496 <wifi> <log> Assert at nxmac_monotonic_counter_2_lo_get:11955     37433 <platform> <feat> watchdog reset at timer_hal_get_timer0_cnt()     37493 <wifi> <perf> if there is no ack for ps null, max retry 8 times in dtim1     41813 <bt> <docs>optimize dut document     41863 <multimedia> <feat> remove unused project in automatic_project list     41908 <pm> <perf>DIGLOD LOW VOLTAGE ENABLE CONFIG(Default enable, some project disable)     41162 <platform> <fix> reading partitions table error     41293 <wifi> <perf> optimize vnd_cal tool,add data check     41724 <platform> <fix> PSRAM drive strength and voltage changed.     41063 <wifi> <fix> When there is a wrong password, connecting to the router development board will not prompt a password error.     41391 <wifi> <fix> expand the bss_max_count based on the number of scan     41934 <bt> <fix> fix adv data buffer num     41575 <platform> <docs> Remove all descriptions about the CM33 architecture and the number of CPU cores in the document     42006 <multimedia> <perf> add lcd example part table     42168 <platform> <fix> Enter the USB shutdown process and directly set the memory pointer to NULL     41990 <platform> <fix> Add initialization completion flag     42195 <doc> adjust mesh docs     42255 <fix> add addr check when set adv random addr     42418 <platform> <perf> Increase gpio docs illustrate     42434 <platform> <perf> Open USB for 480M CPU usage voting

+++ middleware/soc/bk7234/hal/uart_ll.h
@@ -290,7 +290,7 @@ static inline uint8_t *uart_ll_get_write_data_addr(uart_hw_t *hw, uart_id_t id)
 
 static inline uint8_t *uart_ll_get_read_data_addr(uart_hw_t *hw, uart_id_t id)
 {
-	return (uint8_t *)((&hw->fifo_port)+1);
+	return (uint8_t *)(((uint32_t)&hw->fifo_port)+1);
 }
 
 static inline void uart_ll_reset_int_en_to_default(uart_hw_t *hw, uart_id_t id)
